+++ src/c_cmds.c
@@ -6657,7 +6657,7 @@ static void spawn_cmd_func2(char *cmd, char *parms)
                             M_SaveCVARs();
                         }
 
-                        thing->flags3 |= MF3_SPAWNEDBYPLAYER;
+                        thing->flags2 |= MF2_SPAWNEDBYPLAYER;
 
                         if (flags & MF_NOGRAVITY)
                         {