+++ src/d_deh.c
@@ -1519,7 +1519,7 @@ typedef struct
 // killough 08/09/98: make DEH_BLOCKMAX self-adjusting
 #define DEH_BLOCKMAX    arrlen(deh_blocks)              // size of array
 #define DEH_MAXKEYLEN   32                              // as much of any key as we'll look at
-#define DEH_MOBJINFOMAX 42                              // number of mobjinfo configuration keys
+#define DEH_MOBJINFOMAX 41                              // number of mobjinfo configuration keys
 
 // Put all the block header values, and the function to be called when that
 // one is encountered, in this array:
@@ -1587,7 +1587,6 @@ static const char *deh_mobjinfo[DEH_MOBJINFOMAX] =
     "Action sound",             // .activesound
     "Bits",                     // .flags
     "Retro bits",               // .flags2
-    "Retro bits 2",             // .flags3
     "Respawn frame",            // .raisestate
     "Frames",                   // .frames
     "Fullbright",               // .fullbright