+++ src/p_mobj.c
@@ -398,7 +398,7 @@ static void P_ZMovement(mobj_t *mo)
                 }
 
                 // killough 11/98: touchy objects explode on impact
-                if ((flags & MF_TOUCHY) && (mo->flags3 & MF3_ARMED) && mo->health > 0)
+                if ((flags & MF_TOUCHY) && (mo->flags2 & MF2_ARMED) && mo->health > 0)
                     P_DamageMobj(mo, NULL, NULL, mo->health, true);
                 else if ((flags & MF_FLOAT) && sentient(mo))
                     goto floater;
@@ -514,7 +514,7 @@ floater:
         if (mo->momz < 0)
         {
             // killough 11/98: touchy objects explode on impact
-            if ((flags & MF_TOUCHY) && (mo->flags3 & MF3_ARMED) && mo->health > 0)
+            if ((flags & MF_TOUCHY) && (mo->flags2 & MF2_ARMED) && mo->health > 0)
                 P_DamageMobj(mo, NULL, NULL, mo->health, true);
             else if (player && player->mo == mo)
             {
@@ -735,7 +735,7 @@ void P_MobjThinker(mobj_t *mobj)
     }
     else if (!(mobj->momx | mobj->momy) && !sentient(mobj))
     {
-        mobj->flags3 |= MF3_ARMED;  // arm a mine which has come to rest
+        mobj->flags2 |= MF2_ARMED;  // arm a mine which has come to rest
 
         // killough 09/12/98: objects fall off ledges if they are hanging off
         // slightly push off of ledge if hanging more than halfway off
@@ -802,7 +802,6 @@ mobj_t *P_SpawnMobj(fixed_t x, fixed_t y, fixed_t z, mobjtype_t type)
     mobj->radius = info->radius;
     mobj->flags = info->flags;
     mobj->flags2 = info->flags2;
-    mobj->flags3 = info->flags3;
     mobj->mbf21flags = info->mbf21flags;
     mobj->health = info->spawnhealth;
 
@@ -1261,9 +1260,6 @@ mobj_t *P_SpawnMapThing(mapthing_t *mthing, dboolean spawnmonsters)
     // [BH] randomly mirror weapons
     if ((type == SuperShotgun || (type >= Shotgun && type <= BFG9000)) && (M_BigRandom() & 1) && r_mirroredweapons)
         mobj->flags2 |= MF2_MIRRORED;
-
-    if (type == Revenant || type == LostSoul || type == SpiderMastermind || type == Cyberdemon)
-        mobj->flags3 |= MF3_MISSILEMORE;
 
     info = mobj->info;