+++ src/p_saveg.c
@@ -233,7 +233,7 @@ static void saveg_read_mobj_t(mobj_t *str)
     str->state = ((state = saveg_read32()) > 0 && state < NUMSTATES ? &states[state] : NULL);
     str->flags = saveg_read32();
     str->flags2 = saveg_read32();
-    str->flags3 = saveg_read32();
+    str->mbf21flags = saveg_read32();
     str->health = saveg_read32();
     str->movedir = saveg_read32();
     str->movecount = saveg_read32();
@@ -273,9 +273,9 @@ static void saveg_read_mobj_t(mobj_t *str)
     str->madesound = saveg_read32();
     str->inflicter = saveg_read32();
     str->geartime = saveg_read32();
-    str->mbf21flags = saveg_read32();
 
     // [BH] For future features without breaking savegame compatibility
+    saveg_read32();
     saveg_read32();
     saveg_read32();
     saveg_read32();
@@ -313,7 +313,7 @@ static void saveg_write_mobj_t(mobj_t *str)
     saveg_write32((int)(str->state - states));
     saveg_write32(str->flags);
     saveg_write32(str->flags2);
-    saveg_write32(str->flags3);
+    saveg_write32(str->mbf21flags);
     saveg_write32(str->health);
     saveg_write32(str->movedir);
     saveg_write32(str->movecount);
@@ -354,9 +354,9 @@ static void saveg_write_mobj_t(mobj_t *str)
     saveg_write32(str->madesound);
     saveg_write32(str->inflicter);
     saveg_write32(str->geartime);
-    saveg_write32(str->mbf21flags);
 
     // [BH] For future features without breaking savegame compatibility
+    saveg_write32(0);
     saveg_write32(0);
     saveg_write32(0);
     saveg_write32(0);